+++ Source/ACE.Server/Command/Handlers/AdminCommands.cs
@@ -1532,34 +1532,87 @@ namespace ACE.Server.Command.Handlers
         // rename <Current Name> <New Name>
         [CommandHandler("rename", AccessLevel.Envoy, CommandHandlerFlag.None, 2,
             "Rename a character. (Do NOT include +'s for admin names)",
-            "< Current Name > < New Name >")]
+            "< Current Name >, < New Name >")]
         public static void HandleRename(Session session, params string[] parameters)
         {
             // @rename <Current Name>, <New Name> - Rename a character. (Do NOT include +'s for admin names)
 
-            // As currently implemented below, the command does not exactly mimic retail usage in that the command was @rename oldName, newName
-            // and for us is @rename oldName newName || is this a big deal?? -Ripley
-
-            string fixupOldName = "";
-            string fixupNewName = "";
-
-            if (parameters[0] == "" || parameters[1] == "")
-                return;
-
-            fixupOldName = parameters[0].Replace("+", "").Remove(1).ToUpper() + parameters[0].Replace("+", "").Substring(1);
-            fixupNewName = parameters[1].Replace("+", "").Remove(1).ToUpper() + parameters[1].Replace("+", "").Substring(1);
-
-            string message = "";
-
-            DatabaseManager.Shard.RenameCharacter(fixupOldName, fixupNewName, ((uint charId) =>
+            if (!string.Join(" ", parameters).Contains(','))
             {
-                if (charId > 0)
-                    message = $"Character {fixupOldName} has been renamed to {fixupNewName}.";
-                else
-                    message = $"Rename failed because either there is no character by the name {fixupOldName} currently in the database or the name {fixupNewName} is already taken.";
+                CommandHandlerHelper.WriteOutputInfo(session, "Error, cannot rename. You must include the old name followed by a comma and then the new name.\n Example: @rename Old Name, New Name", ChatMessageType.Broadcast);
+                return;
+            }
 
-                CommandHandlerHelper.WriteOutputInfo(session, message, ChatMessageType.WorldBroadcast);
-            }));
+            var names = string.Join(" ", parameters).Split(",");
+
+            var oldName = names[0].TrimStart(' ').TrimEnd(' ');
+            var newName = names[1].TrimStart(' ').TrimEnd(' ');
+
+            if (oldName.StartsWith("+"))
+                oldName = oldName.Substring(1);
+            if (newName.StartsWith("+"))
+                newName = newName.Substring(1);
+
+            newName = newName.First().ToString().ToUpper() + newName.Substring(1);
+
+            var onlinePlayer = PlayerManager.GetOnlinePlayer(oldName);
+            var offlinePlayer = PlayerManager.GetOfflinePlayer(oldName);
+            if (onlinePlayer != null)
+            {
+                var success = false;
+                DatabaseManager.Shard.IsCharacterNameAvailable(newName, isAvailable =>
+                {
+                    if (!isAvailable)
+                    {
+                        CommandHandlerHelper.WriteOutputInfo(session, $"Error, a player named \"{newName}\" already exists.", ChatMessageType.Broadcast);
+                        return;
+                    }
+
+                    onlinePlayer.Character.Name = newName;
+                    onlinePlayer.CharacterChangesDetected = true;
+                    onlinePlayer.Name = newName;
+                    onlinePlayer.SavePlayerToDatabase();
+
+                    success = true;
+                });
+
+                if (success)
+                {
+                    CommandHandlerHelper.WriteOutputInfo(session, $"Player named \"{oldName}\" renamed to \"{newName}\" succesfully!", ChatMessageType.Broadcast);
+
+                    onlinePlayer.Session.LogOffPlayer();
+                }
+            }
+            else if (offlinePlayer != null)
+            {
+                var success = false;
+                DatabaseManager.Shard.IsCharacterNameAvailable(newName, isAvailable =>
+                {
+                    if (!isAvailable)
+                    {
+                        CommandHandlerHelper.WriteOutputInfo(session, $"Error, a player named \"{newName}\" already exists.", ChatMessageType.Broadcast);
+                        return;
+                    }
+
+                    var character = DatabaseManager.Shard.GetCharacterByName(oldName);
+
+                    character.Name = newName;
+                    DatabaseManager.Shard.SaveCharacter(character, new ReaderWriterLockSlim(), null);
+                    offlinePlayer.SetProperty(PropertyString.Name, newName);
+                    offlinePlayer.SaveBiotaToDatabase();
+
+                    success = true;
+                });
+
+                if (success)
+                    CommandHandlerHelper.WriteOutputInfo(session, $"Player named \"{oldName}\" renamed to \"{newName}\" succesfully!", ChatMessageType.Broadcast);
+            }
+            else
+            {
+                CommandHandlerHelper.WriteOutputInfo(session, $"Error, a player named \"{oldName}\" cannot be found.", ChatMessageType.Broadcast);
+                return;
+            }
+
         }
 
         // setadvclass