+++ changelog.md
@@ -4,6 +4,7 @@
 [Ripley]
 * Changed the way Name property is handled with regards to + (Admin/Sentinel characters).
 * Set up basic path to support replicating accesslevel changes on to existing characters.
+* Return @rename to functionality.
 
 ### 2019-03-09
 [OptimShi]